[core] Privatize geojson_tile.hpp contents
